made export stats more flexible
#TODO
Found a bug where some replies are named ['Hide', 'replies'], which must be because in some cases the container has already loaded all replies and the hide button is the only thing visible. Though, this was the thing that kept track of how many replies were in a comment. Maybe needs some sort of counter of replies instead.